disallow deleting home spaces (#306)

+++ src/lib/vocab/space/spaceServices.ts
@@ -21,6 +21,7 @@ import {
 	UpdateSpace,
 	DeleteSpace,
 } from '$lib/vocab/space/spaceEvents';
+import {canDeleteSpace} from '$lib/vocab/space/spaceHelpers';
 import type {Community} from '$lib/vocab/community/community';
 import type {Result} from '@feltcoop/felt';
 import type {Space} from '$lib/vocab/space/space';
@@ -137,6 +138,20 @@ export const deleteSpaceService: Service<DeleteSpaceParams, DeleteSpaceResponseR
 	event: DeleteSpace,
 	perform: async ({repos, params}) => {
 		log.trace('[DeleteSpace] deleting space with id:', params.space_id);
+
+		// Check that the space can be deleted.
+		const findSpaceResult = await repos.space.findById(params.space_id);
+		if (!findSpaceResult.ok) {
+			if (findSpaceResult.type === 'no_space_found') {
+				return {ok: false, status: 404, message: findSpaceResult.message};
+			}
+			return {ok: false, status: 500, message: 'unknown server error'};
+		}
+		const space = findSpaceResult.value;
+		if (!canDeleteSpace(space)) {
+			return {ok: false, status: 405, message: 'cannot delete home space'};
+		}
+
 		const result = await repos.space.deleteById(params.space_id);
 		log.trace('[DeleteSpace] result', result);
 		if (!result.ok) {
